integration of roxanes work
