+++ data/displayState.json
@@ -1,5 +1,5 @@
 {
-	"activeTab": "HOME",
+	"activeTab": "RCS",
 	"activeView": "undefined",
-	"activeRC": 0
+	"activeRC": 1
 }
