+++ data/userPrefs.json
@@ -1,3 +1,3 @@
 {
-    "favoriteRC": 0
+    "favoriteRC": 1
 }
