+++ flux/Utils/displaySchedule.js
