+++ flux/Utils/isItOpen.js
