+++ flux/Utils/isItOpenDay.js
