+++ flux/components/Level.js
