--- src/dateLists.js
@@ -1,6 +1,0 @@
-'use strict';
-
-module.exports = {
-    weekDays : ['lundi', 'mardi', 'mercredi', 'jeudi', 'vendredi', 'samedi', 'dimanche'],
-    months : ["janvier", "février", "mars", "avril", "mai", "juin", "juillet", "août", "septembre", "octobre", "novembre", "décembre"]
-};
