--- src/rcs.json
@@ -1,292 +1,0 @@
-[
-    {
-    
-    //####################################################################################################
-    // NEVER CHANGE
-        "id" : 0,
-
-        "name": "Verac",
-
-        // opened intervall in UTC
-        "schedule": {
-            //UTC
-            //monday
-            "0":  [
-                {"start": "0700", "end": "1000"},
-                {"start": "1200", "end": "1600"}
-            ],
-            // tuesday
-            "1":  [
-                {"start": "0700", "end": "1000"},
-                {"start": "1200", "end": "1600"}
-            ],
-            //wednesday
-            "2":  [
-                {"start": "0700", "end": "1000"},
-                {"start": "1400", "end": "1600"}
-            ],
-            //thursday
-            "3":  [
-                {"start": "0700", "end": "1000"},
-                {"start": "1200", "end": "1600"}
-            ],
-            //friday
-            "4":  [
-                {"start": "0700", "end": "1000"},
-                {"start": "1400", "end": "1600"}
-            ],
-            //saturday
-            "5":  [
-                {"start": "0700", "end": "1000"},
-                {"start": "1200", "end": "1600"}
-            ]
-        },  
-
-        "maxSize": 23,   
-
-        "address": {
-            "number": 5,
-            "typeRoad": "ld",
-            "nameRoad": "Teycheres",
-            "PC": 33240,
-            "town": "Verac"
-        },
-
-        "coords": {
-            "lat": -0.4586,
-            "long": 4.4586
-        },
-
-        "owner": "SMICVAL",
-
-        "phone": "0505050505",
-
-        //#################################################################################################"
-        // CHANGE EACH 5mn
-        "crowd": {
-            //local + GMT (-2 in France)
-
-            "2015-07-27T07:00:00.000Z" : 15,
-
-            "2015-07-27T07:15:00.000Z": 10,
-
-            "2015-07-27T07:30:00.000Z" : 5,
-
-            "2015-07-27T07:45:00.000Z" : 12,
-
-            "2015-07-27T08:00:00.000Z" : 20 ,
-
-            "2015-07-27T08:15:00.000Z" : 15,
-
-            "2015-07-27T08:30:00.000Z" : 18,
-
-            "2015-07-27T08:45:00.000Z" : 25,
-
-            "2015-07-27T09:00:00.000Z" : 30,
-
-            "2015-07-27T09:15:00.000Z" : 20,
-
-            "2015-07-27T09:30:00.000Z" : 16,
-
-            "2015-07-27T09:45:00.000Z" : 15,
-
-            "2015-07-27T10:00:00.000Z" : 15,
-
-            "2015-07-27T12:00:00.000Z" : 14,
-
-            "2015-07-27T12:15:00.000Z" : 13,
-
-            "2015-07-27T12:30:00.000Z" : 10,
-
-            "2015-07-27T12:45:00.000Z" : 8,
-
-            "2015-07-27T13:00:00.000Z" : 16,
-
-            "2015-07-27T13:15:00.000Z" : 18,
-
-            "2015-07-27T13:30:00.000Z" : 15,
-
-            "2015-07-27T13:45:00.000Z" : 13,
-
-            "2015-07-27T14:00:00.000Z" : 10,
-
-            "2015-07-27T14:15:00.000Z" : 19,
-
-            "2015-07-27T14:30:00.000Z" : 22,
-
-            "2015-07-27T14:45:00.000Z" : 18,
-
-            "2015-07-27T15:00:00.000Z" : 15,
-
-            "2015-07-27T15:15:00.000Z" : 20,
-
-            "2015-07-27T15:30:00.000Z" : 25,
-
-            "2015-07-27T15:45:00.000Z" : 15,
-
-            "2015-07-27T16:00:00.000Z" : 15,
-
-    },
-
-        "wastes": 
-        //type never change
-        //status does
-
-        [
-            {
-                "type": "Dechets_verts",
-                "status": "unavailable"
-            },
-            {
-                "type": "Bois",
-                "status": "available"
-            },
-            {
-                "type": "Cartons",
-                "status": "available"
-            }
-        ],
-    },
-    {
-        "id": 1,
-
-        "name": "Verac",
-
-        // opened intervall in UTC
-        "schedule": {
-            //UTC
-            //monday
-            "0":  [
-                {"start": "0700", "end": "1000"},
-                {"start": "1200", "end": "1600"}
-            ],
-            // tuesday
-            "1":  [
-                {"start": "0700", "end": "1000"},
-                {"start": "1200", "end": "1600"}
-            ],
-            "2":  [
-                {"start": "0700", "end": "1000"},
-                {"start": "1200", "end": "1600"}
-            ],
-            "3":  [
-                {"start": "0700", "end": "1000"},
-                {"start": "1200", "end": "1600"}
-            ],
-            "4": [
-                {"start": "0700", "end": "1000"},
-                {"start": "1200", "end": "1600"}
-            ],
-            "5": [
-                {"start": "0700", "end": "1000"},
-                {"start": "1200", "end": "1600"}
-            ]
-        },  
-
-        "maxSize": 23,   
-
-        "address": {
-            "number": 5,
-            "typeRoad": "ld",
-            "nameRoad": "Teycheres",
-            "PC": 33240,
-            "town": "Saint Denis de Pile"
-        },
-
-        "coords": {
-            "lat": -0.4586,
-            "long": 4.4586
-        },
-
-        "owner": "SMICVAL",
-
-        "phone": "0505050505",
-
-        //#################################################################################################"
-        // CHANGE EACH 5mn
-        "crowd": {
-            //local + GMT (-2 in France)
-
-            "2015-07-27T07:00:00.000Z" : 15,
-
-            "2015-07-27T07:15:00.000Z": 10,
-
-            "2015-07-27T07:30:00.000Z" : 5,
-
-            "2015-07-27T07:45:00.000Z" : 12,
-
-            "2015-07-27T08:00:00.000Z" : 20 ,
-
-            "2015-07-27T08:15:00.000Z" : 15,
-
-            "2015-07-27T08:30:00.000Z" : 18,
-
-            "2015-07-27T08:45:00.000Z" : 25,
-
-            "2015-07-27T09:00:00.000Z" : 30,
-
-            "2015-07-27T09:15:00.000Z" : 20,
-
-            "2015-07-27T09:30:00.000Z" : 16,
-
-            "2015-07-27T09:45:00.000Z" : 15,
-
-            "2015-07-27T10:00:00.000Z" : 15,
-
-            "2015-07-27T12:00:00.000Z" : 14,
-
-            "2015-07-27T12:15:00.000Z" : 13,
-
-            "2015-07-27T12:30:00.000Z" : 10,
-
-            "2015-07-27T12:45:00.000Z" : 8,
-
-            "2015-07-27T13:00:00.000Z" : 16,
-
-            "2015-07-27T11:13:00.000Z" : 18,
-
-            "2015-07-27T13:30:00.000Z" : 15,
-
-            "2015-07-27T13:45:00.000Z" : 13,
-
-            "2015-07-27T14:00:00.000Z" : 10,
-
-            "2015-07-27T14:15:00.000Z" : 19,
-
-            "2015-07-27T14:30:00.000Z" : 22,
-
-            "2015-07-27T14:45:00.000Z" : 18,
-
-            "2015-07-27T15:00:00.000Z" : 15,
-
-            "2015-07-27T15:15:00.000Z" : 20,
-
-            "2015-07-27T15:30:00.000Z" : 25,
-
-            "2015-07-27T15:45:00.000Z" : 15,
-
-            "2015-07-27T16:00:00.000Z" : 15,
-
-    },
-
-        "wastes": 
-        //type never change
-        //status does
-
-        [
-            {
-                "type": "Dechets_verts",
-                "status": "unavailable"
-            },
-            {
-                "type": "Bois",
-                "status": "available"
-            },
-            {
-                "type": "Cartons",
-                "status": "available"
-            }
-        ],
-    }
-]
-
