--- src/userFake.json
@@ -1,3 +1,0 @@
-{
-    "favouriteRC": 0
-}
